Changed from float to double to handle higher score values.

+++ src/expectiminimax.cpp
@@ -12,24 +12,24 @@ moves::type exp::bestMove(uint64_t board) {
   if (game::gameOver(board))
     return moves::type::NONE;
 
-  std::unordered_map<uint64_t, float> table;
+  std::unordered_map<uint64_t, double> table;
 
   uint64_t left = moves::moveLeft(board);
   uint64_t right = moves::moveRight(board);
   uint64_t down = moves::moveDown(board);
   uint64_t up = moves::moveUp(board);
 
-  float left_score = (left ^ board) ? exp::chanceNode(left, exp::MAX_DEPTH, table, 1) : 0;
-  float right_score = (right ^ board) ? exp::chanceNode(right, exp::MAX_DEPTH, table, 1) : 0;
-  float up_score = (up ^ board) ? exp::chanceNode(up, exp::MAX_DEPTH, table, 1) : 0;
-  float down_score = (down ^ board) ? exp::chanceNode(down, exp::MAX_DEPTH, table, 1) : 0;
+  double left_score = (left ^ board) ? exp::chanceNode(left, exp::MAX_DEPTH, table, 1) : 0;
+  double right_score = (right ^ board) ? exp::chanceNode(right, exp::MAX_DEPTH, table, 1) : 0;
+  double up_score = (up ^ board) ? exp::chanceNode(up, exp::MAX_DEPTH, table, 1) : 0;
+  double down_score = (down ^ board) ? exp::chanceNode(down, exp::MAX_DEPTH, table, 1) : 0;
 
   // std::cout << "Left Score: " << left_score << "\n";
   // std::cout << "Right Score: " << right_score << "\n";
   // std::cout << "Down Score: " << down_score << "\n";
   // std::cout << "Up Score: " << up_score << "\n\n";
 
-  float best_score = up_score;
+  double best_score = up_score;
   moves::type best_move = moves::type::UP;
 
   if (right_score > best_score) {
@@ -64,7 +64,7 @@ moves::type exp::bestMove(uint64_t board) {
   return best_move;
 }
 
-float exp::stateNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t, float> &table, float probability) {
+double exp::stateNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t, double> &table, double probability) {
   if (table.find(board) != table.end()) {
     return table[board];
   }
@@ -80,12 +80,12 @@ float exp::stateNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t,
   uint64_t down = moves::moveDown(board);
   uint64_t up = moves::moveUp(board);
 
-  float left_score = (left ^ board) ? exp::chanceNode(left, depth - 1, table, probability) : 0.f;
-  float right_score = (right ^ board) ? exp::chanceNode(right, depth - 1, table, probability) : 0.f;
-  float up_score = (up ^ board) ? exp::chanceNode(up, depth - 1, table, probability) : 0.f;
-  float down_score = (down ^ board) ? exp::chanceNode(down, depth - 1, table, probability) : 0.f;
+  double left_score = (left ^ board) ? exp::chanceNode(left, depth - 1, table, probability) : 0.f;
+  double right_score = (right ^ board) ? exp::chanceNode(right, depth - 1, table, probability) : 0.f;
+  double up_score = (up ^ board) ? exp::chanceNode(up, depth - 1, table, probability) : 0.f;
+  double down_score = (down ^ board) ? exp::chanceNode(down, depth - 1, table, probability) : 0.f;
 
-  float best_score = std::max(std::max(left_score, right_score), std::max(up_score, down_score));
+  double best_score = std::max(std::max(left_score, right_score), std::max(up_score, down_score));
 
   for (uint64_t transition : bitboard::allFlips(board))
     table[transition] = best_score;
@@ -93,20 +93,20 @@ float exp::stateNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t,
   return best_score;
 }
 
-float exp::chanceNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t, float> &table, float probability) {
+double exp::chanceNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t, double> &table, double probability) {
   uint64_t open_spaces = game::openSpaces(board);
 
   if (!(open_spaces))
     return exp::stateNode(board, depth, table, probability);
 
   uint8_t num_children{game::countTiles(open_spaces)};
-  float prob_two{0.9f * (1.0f / num_children)};
-  float prob_four{0.1f * (1.0f / num_children)};
+  double prob_two{0.9f * (1.0f / num_children)};
+  double prob_four{0.1f * (1.0f / num_children)};
 
   uint64_t new_two = 0x1;
   uint64_t new_four = 0x2;
 
-  float value{0};
+  double value{0};
 
   while (open_spaces) {
     if (open_spaces & 0xF) {
