+++ src/expectiminimax.hpp
@@ -9,11 +9,11 @@
 
 namespace exp {
   constexpr uint8_t MAX_DEPTH = 8;
-  constexpr float PROB_THRESH = 0.0001f;
+  constexpr double PROB_THRESH = 0.0001f;
 
-  float stateNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t, float> &table, float probability);
+  double stateNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t, double> &table, double probability);
 
-  float chanceNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t, float> &table, float probability);
+  double chanceNode(uint64_t board, uint8_t depth, std::unordered_map<uint64_t, double> &table, double probability);
 
   moves::type bestMove(uint64_t board);
 }  // namespace exp
